fix: updated baseUrl of links in footer, move data sources into adata folder
