+++ src/components/Audience.astro
@@ -1,5 +1,5 @@
 ---
-import { audienceTypes } from "../audience";
+import { audienceTypes } from "../data/audience";
 
 const { audience = [] } = audienceTypes;
 const { heading } = Astro.props;
