+++ src/components/Hero.astro
@@ -1,6 +1,6 @@
 ---
 import { Icon } from "astro-icon/components";
-import { heroItems } from "../navigation";
+import { heroItems } from "../data/navigation";
 import GradientBanner from "./GradientBanner.astro";
 const { links = [] } = heroItems;
 const base = import.meta.env.BASE_URL;
