+++ src/components/QuickAccessShortcuts.astro
@@ -1,5 +1,5 @@
 ---
-import { quickAccessShortcuts } from "../navigation";
+import { quickAccessShortcuts } from "../data/navigation";
 ---
 
 <div class="container-fluid pt-10 pb-16 mx-8 px-6 lg:px-10 lg:mx-10 xl:px-16 xl:mx-24 bg-transparentTeal">
