+++ src/data/navigation.ts
@@ -1,3 +1,6 @@
+const base = import.meta.env.BASE_URL;
+
+
 export const headerData = {
   links: [
     {
@@ -263,11 +266,11 @@ export const footerData = {
   links: [
     { text: 'SuperOffice', href: 'https://www.superoffice.com/' },
     { text: 'Community', href: 'https://community.superoffice.com' },
-    { text: 'Release Notes', href: '/release-notes' },
+    { text: 'Release Notes', href: `${base}/release-notes` },
     { text: 'Privacy', href: 'https://www.superoffice.com/company/privacy/' },
     { text: 'Site feedback', href: 'https://github.com/SuperOfficeDocs/feedback/issues/new?title=Feedback%20for%20SuperOffice%20Docs&body=%0A%0A%5BEnter%20feedback%20here%5D%0A%0A%0A---%0A%23%23%23%23%20Document%20Details%0A%0A%E2%9A%A0%20*Do%20not%20edit%20this%20section.%20It%20is%20required%20for%20docs.superOffice.com%20%E2%9E%9F%20Docs%20Team%20processing.*%0A%0A*%20Content%20Source%3A%20%5BSuperOffice%20Docs%5D(https%3A%2F%2Fdocs.superoffice.com)'},
-    { text: 'About', href: '/about' },
-    { text: 'Contribute', href:'/contribute/overview'},
+    { text: 'About', href: `${base}/about` },
+    { text: 'Contribute', href:`${base}/contribute/overview`},
     { text: 'Back to top', href:'#'}
   ],
 };
