+++ src/layouts/Base.astro
@@ -2,7 +2,7 @@
 import favIcon from "@assets/favicon.ico";
 import Header from "@components/Header.astro";
 import Footer from "@components/Footer.tsx";
-import { footerData, headerData } from "../navigation";
+import { footerData, headerData } from "../data/navigation";
 import { ClientRouter } from "astro:transitions";
 import "@styles/main.css";
