+++ src/pages/index.astro
@@ -4,7 +4,7 @@ import AreasOfInterest from "@components/AreasOfInterest.astro";
 import Audience from "@components/Audience.astro";
 import QuickAccessShortcuts from "@components/QuickAccessShortcuts.astro";
 import Layout from "@layouts/Base.astro";
-import { areasOfInterest } from "../navigation";
+import { areasOfInterest } from "../data/navigation";
 
 const metadata = {
   title: "SuperOffice Docs",
